--- apps/astro-party/src/lib/castingwords.test.ts
@@ -1,3 +1,0 @@
-import {buildCastingwordsOrderUrl, orderTranscript} from './castingwords'
-
-console.log('snarf')
