+++ apps/scriptkit/src/pages/api/webhooks/sanity/videoResource/created.ts
@@ -2,7 +2,7 @@ import {withSentry} from '@sentry/nextjs'
 import {NextApiRequest, NextApiResponse} from 'next'
 import {isValidSignature, SIGNATURE_HEADER_NAME} from '@sanity/webhook'
 import * as Sentry from '@sentry/nextjs'
-import {createCastingWordsOrder} from '@skillrecordings/skill-lesson/lib/casting-words'
+import {createCastingWordsOrder} from '@skillrecordings/skill-lesson/lib/castingwords'
 import {createMuxAsset} from '@skillrecordings/skill-lesson/lib/mux'
 import {updateVideoResourceWithTranscriptOrderId} from '@skillrecordings/skill-lesson/lib/sanity'
